+++ src/MC_C_ISO_TUN.cpp
@@ -12,7 +12,7 @@
 using namespace Rcpp;
 
 // [[Rcpp::export("MC_C_ISO_TUN")]]
-List MC_C_ISO_TUN(arma::vec times, int N_e, arma::vec r, double rho, double E, double s, double T) {
+List MC_C_ISO_TUN(arma::vec times, double N_e, arma::vec r, double rho, double E, double s, double T) {
 
   //determine delta_t which allows to have delta t != 1
   double delta_t = calc_deltat(times);
@@ -29,7 +29,7 @@ List MC_C_ISO_TUN(arma::vec times, int N_e, arma::vec r, double rho, double E, d
 
       std::size_t n_filled = N_e;
 
-      double P = ((s * exp(-E/(k_B * (273 + T)))) * exp(-(pow(rho,-1.0/3)) * r[k]));
+      double P = ((s * exp(-E/(k_B * (273 + T)))) * exp(-(pow(rho,-1.0/3.0)) * r[k]));
 
       for(std::size_t t = 0; t < times.size(); ++t){
         for(std::size_t j = 0; j < n_filled; ++j){
@@ -44,7 +44,7 @@ List MC_C_ISO_TUN(arma::vec times, int N_e, arma::vec r, double rho, double E, d
             break;
 
         } // end n_filled
-        signal(t,k) = n_filled * P * 3 * pow((double)r[k],2) * exp(-(pow(r[k],3)));
+        signal(t,k) = n_filled * P * 3 * pow((double)r[k],2.0) * exp(-(pow(r[k],3.0)));
         remaining_e(t,k) = n_filled;
 
         if (n_filled == 0)
